Moved observers to separate package
