+++ app/src/main/java/io/lunarlogic/aircasting/models/observers/DormantSessionsObserver.kt
@@ -1,7 +1,9 @@
-package io.lunarlogic.aircasting.models
+package io.lunarlogic.aircasting.models.observers
 
 import androidx.lifecycle.LifecycleOwner
 import io.lunarlogic.aircasting.database.data_classes.SessionWithStreamsDBObject
+import io.lunarlogic.aircasting.models.Session
+import io.lunarlogic.aircasting.models.SessionsViewModel
 import io.lunarlogic.aircasting.screens.dashboard.SessionsViewMvc
 
 class DormantSessionsObserver(
