+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/SessionsController.kt
@@ -2,9 +2,7 @@ package io.lunarlogic.aircasting.screens.dashboard
 
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.LiveData
 import io.lunarlogic.aircasting.database.DatabaseProvider
-import io.lunarlogic.aircasting.database.data_classes.SessionWithStreamsAndMeasurementsDBObject
 import io.lunarlogic.aircasting.screens.new_session.NewSessionActivity
 import io.lunarlogic.aircasting.exceptions.ErrorHandler
 import io.lunarlogic.aircasting.lib.NavigationController
@@ -15,7 +13,6 @@ import io.lunarlogic.aircasting.networking.services.SessionsSyncService
 import io.lunarlogic.aircasting.screens.session_view.graph.GraphActivity
 import io.lunarlogic.aircasting.screens.session_view.map.MapActivity
 import io.lunarlogic.aircasting.models.Session
-import io.lunarlogic.aircasting.models.SessionsObserver
 import io.lunarlogic.aircasting.models.SessionsViewModel
