+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/active/MobileActiveController.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.LifecycleOwner
 import io.lunarlogic.aircasting.events.StopRecordingEvent
 import io.lunarlogic.aircasting.lib.NavigationController
 import io.lunarlogic.aircasting.lib.Settings
-import io.lunarlogic.aircasting.models.ActiveSessionsObserver
+import io.lunarlogic.aircasting.models.observers.ActiveSessionsObserver
 import io.lunarlogic.aircasting.models.Session
 import io.lunarlogic.aircasting.models.SessionsViewModel
 import io.lunarlogic.aircasting.screens.dashboard.DashboardPagerAdapter
