+++ app/src/main/java/io/lunarlogic/aircasting/screens/session_view/SessionDetailsViewController.kt
@@ -5,6 +5,7 @@ import io.lunarlogic.aircasting.database.DatabaseProvider
 import io.lunarlogic.aircasting.events.NewMeasurementEvent
 import io.lunarlogic.aircasting.location.LocationHelper
 import io.lunarlogic.aircasting.models.*
+import io.lunarlogic.aircasting.models.observers.SessionObserver
 import io.lunarlogic.aircasting.screens.session_view.hlu.HLUValidationErrorToast
 import io.lunarlogic.aircasting.screens.dashboard.SessionPresenter
 import kotlinx.coroutines.CoroutineScope
